Logd socket reconnects
Under some unknown circumstances, writes to the logd socket fail. The
android liblog implementation reacts on such errors with retries and
reconnects. The android-logd-logger expected every write to a once
opened socket to be ok which is not feasible in real world applications.

This patch removes the expectation that a write to an open logd socket
succeeds. Upon a failed write, a new socket is connected and the write
attempt is repeated once. If the second connect or write attempt fails a
message is logged to stderr.

Remove the tls feature.