Use real nanoseconds instead of sequence number
The C version of the pmsg writer states that it hijacks the nanoseconds
field to store a sequence number in microseconds. This seems to be
right: In the code, the sequence number counts up from zero in
increments of 1000 (1000ns = 1µs). By storing the sequence number in the
nanoseconds field, the timestamps of `logd` and `pstore` no longer
match. Other logs (generated from Java code) have matching timestamps
(also in the nanoseconds).

We tried writing the actual nanoseconds and dropping the sequence number
and could not observe any adverse effect. The timestamps now match
between `logd` and `pstore`.