Remove unnecessary arguments for log(...)
We don't need to specify PID in the public log(...) function
because logd obtains it from the credentials ([0]), which
we don't fill explicitly (no stable api in Rust for that now).
However, in our case (logdw socket has enabled passcred option)
the kernel does it automatically ([1]).
We also shouldn't pass TID (in public log(...)) because
we always use our thread id for that purpose.

 [0] https://chromium.googlesource.com/aosp/platform/system/logging/+/refs/heads/factory-ambassador-14265.B/logd/LogListener.cpp#119
 [1] https://elixir.bootlin.com/linux/v6.5/source/net/unix/af_unix.c#L1857