fix: Enable lowering log level below startup value (#15)
`Logger::parse_filters` allows to change the log filter even after the
logger has been initialized. However, it was not possible to see logs
which were above the level set at startup. E.g. if the logger was
initialized with `debug`, setting it to `trace` at runtime did not work
since the filter of the global logging facade filtered out the `trace`
logs.

This commit fixes the issue by calling `log::set_max_level` with the
level from the updated log filter.

Co-authored-by: Simon B. Gasse <sgasse@users.noreply.github.com>